Fixing bigint column types.

+++ db/migrate/20120104212209_create_stars.rb
@@ -4,7 +4,7 @@ class CreateStars < ActiveRecord::Migration
       t.string :name
       t.string :classification
       t.integer :radius
-      t.integer :mass
+      t.bigint :mass
       t.integer :orbit
       t.integer :system_id
